admin_service/app/requirements.txt
django==4.0.4
psycopg2-binary==2.9.6
uwsgi==2.0.20
flake8==4.0
python-dotenv==1.0.0
django-split-settings==1.2.0
django-debug-toolbar==3.4.0
isort==5.12.0
djangorestframework==3.14.0
requests==2.32.3
PyJWT==2.8.0